chore: modernize code
