+++ internal/utils/testutils/socket_windows.go
@@ -1,0 +1,11 @@
+//go:build windows
+
+package testutils
+
+func GetGIDOfFile(_ string) (int, error) {
+	return 0, nil
+}
+
+func GetPermissionsOfFile(_ string) (string, error) {
+	return "", nil
+}
